allowed mutation operation only for admin users

+++ src/services/apiBookings.js
@@ -106,7 +106,9 @@ export async function getStaysTodayActivity() {
   return data;
 }
 
-export async function updateBooking(id, obj) {
+export async function updateBooking(id, obj, isAdmin) {
+  if (!isAdmin) throw new Error('Only admins can do this action');
+
   const { data, error } = await supabase
     .from('bookings')
     .update(obj)
@@ -121,7 +123,9 @@ export async function updateBooking(id, obj) {
   return data;
 }
 
-export async function deleteBooking(id) {
+export async function deleteBooking(id, isAdmin) {
+  if (!isAdmin) throw new Error('Only admins can do this action');
+
   // REMEMBER RLS POLICIES
   const { data, error } = await supabase.from('bookings').delete().eq('id', id);
